chore: move `with-manager.ts` to database folder
