+++ test/seed/catalogue/container-seeder.ts
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import WithManager from '../../../src/with-manager';
+import WithManager from '../../../src/database/with-manager';
 import User, { UserType } from '../../../src/entity/user/user';
 import Container from '../../../src/entity/container/container';
 import ProductRevision from '../../../src/entity/product/product-revision';
