+++ test/seed/catalogue/product-category-seeder.ts
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import WithManager from '../../../src/with-manager';
+import WithManager from '../../../src/database/with-manager';
 import ProductCategory from '../../../src/entity/product/product-category';
 
 export default class ProductCategorySeeder extends WithManager {
