+++ test/seed/catalogue/vat-group-seeder.ts
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import WithManager from '../../../src/with-manager';
+import WithManager from '../../../src/database/with-manager';
 import VatGroup from '../../../src/entity/vat-group';
 import { VatGroupRequest } from '../../../src/controller/request/vat-group-request';
