+++ test/seed/event-seeder.ts
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import WithManager from '../../src/with-manager';
+import WithManager from '../../src/database/with-manager';
 import Role from '../../src/entity/rbac/role';
 import EventShift from '../../src/entity/event/event-shift';
 import User from '../../src/entity/user/user';
