+++ test/seed/ledger/transaction-seeder.ts
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import WithManager from '../../../src/with-manager';
+import WithManager from '../../../src/database/with-manager';
 import PointOfSaleRevision from '../../../src/entity/point-of-sale/point-of-sale-revision';
 import User from '../../../src/entity/user/user';
 import Transaction from '../../../src/entity/transactions/transaction';
