+++ test/seed/user-seeder.ts
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import WithManager from '../../src/with-manager';
+import WithManager from '../../src/database/with-manager';
 import User, { TermsOfServiceStatus, UserType } from '../../src/entity/user/user';
 import InvoiceUser from '../../src/entity/user/invoice-user';
 import bcrypt from 'bcrypt';
